migrate tag component to parents of Tag
