+++ packages/client/src/pages/Main/containers/EntityDetailBox/EntityDetail/EntityDetailUsedInTable/EntityDetailStatementsTable/EntityDetailStatementsTable.tsx
@@ -47,7 +47,7 @@ export const EntityDetailStatementsTable: React.FC<
           return (
             <>
               {entity && (
-                <EntityTag key={entity.id} entity={entity} showOnly="entity" />
+                <EntityTag key={entity.id} entity={entity} showOnly="tag" />
               )}
             </>
           );
