+++ packages/client/src/pages/Main/containers/StatementsListBox/StatementListTable/StatementListTable.tsx
@@ -258,7 +258,7 @@ export const StatementListTable: React.FC<StatementListTable> = ({
         Header: "",
         Cell: ({ row }: CellType) => {
           const statement = row.original;
-          return <EntityTag entity={statement} showOnly="entity" />;
+          return <EntityTag entity={statement} showOnly="tag" />;
         },
       },
       {
